Create ClaimAccountButton component.
