+++ js/src/components/google-ads-account-card/claim-account/index.scss
@@ -1,6 +1,5 @@
 .gla-ads-claim-account-notice {
 	background-color: #ffeec1;
-	border: 0;
 	font-size: $helptext-font-size;
 	margin: 0 var(--large-gap) var(--large-gap);
 	padding: $grid-unit-20;
